Faz alguns ajustes de layout e remove o Ellipse do Header
