+++ NSBrazilConf/Views/Home/Feed/DateFeedView.swift
@@ -26,8 +26,7 @@ struct DateFeedView: View, FeedViewProtocol {
         .frame(maxWidth: .infinity, minHeight: 88)
         .background(Color.white)
         .cornerRadius(4)
-        .padding(.leading, 24)
-        .padding(.trailing, 24)
+        .padding(.horizontal, 10)
         .shadow(color: Color.gray.opacity(0.3), radius: 8, x: 0, y: 6)
     }
