+++ NSBrazilConf/Views/Home/Feed/MapFeedView.swift
@@ -30,7 +30,6 @@ struct MapFeedView: View, FeedViewProtocol {
         .cornerRadius(4)
         .frame(maxWidth: .infinity, minHeight: 286)
         .shadow(color: Color.gray.opacity(0.4), radius: 8, x: 0, y: 6)
-        .padding(.leading, 24)
-        .padding(.trailing, 24)
+        .padding(.horizontal, 10)
     }
 }
