+++ NSBrazilConf/Views/Home/Feed/TitleHeaderView.swift
@@ -3,11 +3,11 @@ import SwiftUI
 
 struct TitleHeaderView: View {
     init?(feedItem: FeedItem) {
-           guard let item = feedItem as? SubtitleFeedItem else { return nil }
-           self.feedItem = item
-       }
+        guard let item = feedItem as? SubtitleFeedItem else { return nil }
+        self.feedItem = item
+    }
 
-       var feedItem: SubtitleFeedItem
+    var feedItem: SubtitleFeedItem
     
     var body: some View {
         ZStack(alignment: .leading) {
@@ -22,8 +22,7 @@ struct TitleHeaderView: View {
             }
         }
         .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
-        .padding(.leading, 24)
-
+        .padding(.horizontal, 10)
     }
 
 }
